Create proper main file
In a previous commit, the main file was renamed to main_tests.c, since
its job was more to test if the new sorting algorithm was working
properly (using bash's file redirectioning) than to serve as a final
main function. This commit creates this new main.

It was a little bit of work, it took a lot of study to know how to deal
with functions pointer, but now it is working just right.

This function must read all the month files, sort it with all different
sorting algorithms and generates a new sorted file and a table file
(using free format csv) for recording stats about how many moves and
comparisons it made.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>