Adapt sorting algorithsm to work with current main
Before, sorting functions would return the sorted array of strings. But
this was just to make sure that the array was sorted. But it is more
important, with main function using pointers to function, to know what
function is being executed, so, as of now, each function return its name
as a string, so it can be manipulated.

Also removes included library, so it doesn't gives any redefinition
error.

TODO: change file extension from .h to .c?

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>