Finished version 1 of Python file
I did some minor modifications in the Python file, using os lib to get
all csv files generated and using them to create png files of the graph.

I did a little research and found out that to generate the image just
like using show function, you can get dpi of figure and then setting the
savefig dpi to it. It worked. I also added the name of the sorting
algorithm to the graph title.

I expect revision in this commit.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>