Correct file checking in bash script
Re-structure output folder, crating a folder for each month, so each
file needs to be named only with the algorithm name. Also change the
name of the bash folder to etc.
Change the compare script to get the files in the folder instead of
needing to write in the code their names, using the new naming pattern
to print messages about all missorted files.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>