Creates a new Bash script to test sorting methods
This script checks two things in the sorting methods: if the output file
is correct ordered, by using two simple test cases, and if the output
data file has the correct numbers, comparing with the numbers given in
the content by the professor.

It was needed to change one function from the sorting main.h (the
function to write the csv file), also adapting it to receive the path of
the output file instead of just the name, so now it can be used to
create files in different paths.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>
Co-authored-by> Pauloguilhermepp <pauloguilherme@usp.br>