Refactored run_tests script and added to run script
Added the run_tests script in the format script and executed it so it
could be correctly formatted.

Also, apperently was shake_sort the algorithm to have its expected
number wrong in one of the files: it was having a wrong answer only in
test2.txt file. The expected value was 64 and gotten value was 84,
meaning, similar numbers.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>