Improve heap_sort algorithm
It is still not working just like Renato's. But I am each time more
convinced that the numbers in the PDF are wrong. The algorithm is
working just fine and I can't think of any more precise way to measure
its comparisons and moves and yet the numbers are divergent.

Signed-off-by: lucascardoso910 <lucascardoso910@gmail.com>